minor TestSettings class refactor
